Done two core tables CRUD. Added protected routes.

+++ src/app/components/medical-records/medical-records.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { MedicalRecord } from 'src/app/interfaces/medical-record';
 import { MedicalRecordService } from 'src/app/services/medical-record.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-medical-records',
@@ -12,6 +13,8 @@ export class MedicalRecordsComponent implements OnInit {
 
   constructor(private medicalRecordService: MedicalRecordService) {}
   ngOnInit(): void {
+    console.log('ON INIT del medical records component');
+    console.log(this.getMedicalRecords());
     this.getMedicalRecords();
   }
 
@@ -23,12 +26,29 @@ export class MedicalRecordsComponent implements OnInit {
   }
 
   deleteMedicalRecord(id: number | undefined) {
-    this.medicalRecordService.deleteMedicalRecord(id).subscribe(
-      (res) => {
-        console.log(res);
-        this.getMedicalRecords();
-      },
-      (err) => console.log(err)
-    );
+    Swal.fire({
+      title: 'Está seguro de eliminar?',
+      text: 'No podrá revertir esta acción',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminar historia clínica',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Se ha eliminado la historia clínica',
+          'Historia clínica eliminada',
+          'success'
+        );
+        this.medicalRecordService.deleteMedicalRecord(id).subscribe(
+          (res) => {
+            console.log(res);
+            this.getMedicalRecords();
+          },
+          (err) => console.log(err)
+        );
+      }
+    });
   }
 }
